# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Geordi Platform API
  version: milestone-3-local
  description: Platform APIs and the locally implemented Metrics and Traces capabilities, pending GitLab revalidation.
servers:
  - url: http://localhost:8080
paths:
  /api/platform:
    get:
      operationId: getPlatform
      summary: Return Geordi platform identity.
      responses:
        '200':
          description: Platform identity.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Platform'
  /api/modules:
    get:
      operationId: getModules
      summary: Return all installed compile-time modules ordered by id.
      responses:
        '200':
          description: Registered modules, including disabled modules.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Modules'
  /api/platform/health:
    get:
      operationId: getPlatformHealth
      summary: Return aggregate platform and module health.
      description: >-
        Always returns HTTP 200 when the backend can answer, even when the reported
        status is DOWN. Disabled modules are not health-checked and do not degrade
        aggregate status.
      responses:
        '200':
          description: Aggregate platform health and module health snapshots.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlatformHealth'
  /api/metrics/services:
    get:
      operationId: listMonitoredServices
      summary: List monitored services with metrics in a bounded time range.
      description: Only explicitly classified monitored workload telemetry is returned.
      parameters:
        - $ref: '#/components/parameters/From'
        - $ref: '#/components/parameters/To'
      responses:
        '200':
          description: Services ordered by their composite OpenTelemetry identity.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MonitoredServices'
        '400':
          $ref: '#/components/responses/BadRequest'
        '503':
          $ref: '#/components/responses/MetricsUnavailable'
        '404':
          $ref: '#/components/responses/MetricsDisabled'
  /api/metrics/overview:
    get:
      operationId: getMetricsOverview
      summary: Return the fixed operational metrics overview for one service.
      parameters:
        - $ref: '#/components/parameters/ServiceName'
        - $ref: '#/components/parameters/ServiceNamespace'
        - $ref: '#/components/parameters/Environment'
        - $ref: '#/components/parameters/From'
        - $ref: '#/components/parameters/To'
      responses:
        '200':
          description: Latest available values for the supported operational metrics.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MetricsOverview'
        '400':
          $ref: '#/components/responses/BadRequest'
        '503':
          $ref: '#/components/responses/MetricsUnavailable'
        '404':
          $ref: '#/components/responses/MetricsDisabled'
  /api/metrics/series:
    get:
      operationId: queryMetricSeries
      summary: Query one or more supported operational metric series.
      description: Resolution is selected by Geordi; backend query languages are never accepted.
      parameters:
        - $ref: '#/components/parameters/ServiceName'
        - $ref: '#/components/parameters/ServiceNamespace'
        - $ref: '#/components/parameters/Environment'
        - $ref: '#/components/parameters/From'
        - $ref: '#/components/parameters/To'
        - name: metric
          in: query
          required: true
          description: Repeat this parameter to request several supported series.
          schema:
            type: array
            minItems: 1
            uniqueItems: true
            items:
              $ref: '#/components/schemas/OperationalMetric'
          explode: true
      responses:
        '200':
          description: Ordered requested series; missing telemetry produces an empty points array.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MetricSeriesResult'
        '400':
          $ref: '#/components/responses/BadRequest'
        '503':
          $ref: '#/components/responses/MetricsUnavailable'
        '404':
          $ref: '#/components/responses/MetricsDisabled'
  /api/traces/services:
    get:
      operationId: listTraceServices
      summary: List monitored services with traces in a bounded half-open time range.
      description: >-
        Results include only explicitly monitored traces
        and are ordered by exact service namespace/name/environment identity.
      parameters:
        - $ref: '#/components/parameters/From'
        - $ref: '#/components/parameters/To'
      responses:
        '200':
          description: Services with monitored traces in the requested [from, to) interval.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TraceServices'
        '400':
          $ref: '#/components/responses/BadRequest'
        '503':
          $ref: '#/components/responses/TracesUnavailable'
        '502':
          $ref: '#/components/responses/TracesMalformedResponse'
        '504':
          $ref: '#/components/responses/TracesTimeout'
        '404':
          $ref: '#/components/responses/TracesDisabled'
  /api/traces:
    get:
      operationId: searchTraces
      summary: Search trace summaries for one exact monitored service identity.
      description: >-
        Search is constrained to [from, to), at most six
        hours and 50 summaries. TraceQL, arbitrary attributes and provider filters are
        never accepted.
      parameters:
        - $ref: '#/components/parameters/TraceServiceName'
        - $ref: '#/components/parameters/TraceServiceNamespace'
        - $ref: '#/components/parameters/TraceEnvironment'
        - $ref: '#/components/parameters/From'
        - $ref: '#/components/parameters/To'
        - name: errorOnly
          in: query
          required: false
          schema:
            type: boolean
            default: false
      responses:
        '200':
          description: Deterministically ordered trace summaries, up to 50.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TraceSearchResult'
        '400':
          $ref: '#/components/responses/BadRequest'
        '503':
          $ref: '#/components/responses/TracesUnavailable'
        '502':
          $ref: '#/components/responses/TracesMalformedResponse'
        '504':
          $ref: '#/components/responses/TracesTimeout'
        '404':
          $ref: '#/components/responses/TracesDisabled'
  /api/traces/{traceId}:
    get:
      operationId: getTrace
      summary: Return complete available span data for one trace.
      description: Provider-specific payloads are never returned.
      parameters:
        - name: traceId
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/TraceId'
      responses:
        '200':
          description: Trace detail with complete available span data and deterministic hierarchy ordering.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TraceDetail'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/TraceNotFound'
        '503':
          $ref: '#/components/responses/TracesUnavailable'
        '502':
          $ref: '#/components/responses/TracesMalformedResponse'
        '504':
          $ref: '#/components/responses/TracesTimeout'
components:
  parameters:
    ServiceName:
      name: serviceName
      in: query
      required: true
      schema:
        type: string
        minLength: 1
    ServiceNamespace:
      name: serviceNamespace
      in: query
      required: false
      schema:
        type: string
        minLength: 1
    Environment:
      name: environment
      in: query
      required: true
      schema:
        type: string
        minLength: 1
    TraceServiceName:
      name: serviceName
      in: query
      required: true
      schema:
        type: string
        minLength: 1
    TraceServiceNamespace:
      name: serviceNamespace
      in: query
      required: false
      description: >-
        When omitted, matches only traces with a canonical null/absent service namespace;
        it never widens the search to every namespace.
      schema:
        type: string
        minLength: 1
    TraceEnvironment:
      name: environment
      in: query
      required: true
      schema:
        type: string
        minLength: 1
    From:
      name: from
      in: query
      required: true
      schema:
        type: string
        format: date-time
    To:
      name: to
      in: query
      required: true
      schema:
        type: string
        format: date-time
  responses:
    BadRequest:
      description: The service identity, range, or requested metric is invalid.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    MetricsUnavailable:
      description: Metrics storage could not answer the bounded query.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    MetricsDisabled:
      description: The Metrics capability is disabled and its routes are absent.
    TracesUnavailable:
      description: Trace storage could not answer the bounded query.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    TracesMalformedResponse:
      description: Trace storage returned a malformed or partial response.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    TracesTimeout:
      description: Trace storage exceeded the bounded query timeout.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    TracesDisabled:
      description: The Traces capability is disabled and its routes are absent.
    TraceNotFound:
      description: No trace exists for the valid requested trace ID.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
  schemas:
    Platform:
      type: object
      additionalProperties: false
      required: [id, name, version]
      properties:
        id:
          type: string
          const: geordi
        name:
          type: string
          const: Geordi
        version:
          type: string
          description: Application version derived from Maven build metadata.
          example: 0.1.0-SNAPSHOT
    ModuleStatus:
      type: string
      enum: [UP, DOWN, UNKNOWN, DISABLED]
    ModuleInventory:
      type: object
      additionalProperties: false
      required: [id, name, enabled]
      properties:
        id:
          type: string
          example: self-observability
        name:
          type: string
          example: Self Observability
        enabled:
          type: boolean
    ModuleHealth:
      type: object
      additionalProperties: false
      required: [id, name, enabled, status]
      properties:
        id:
          type: string
          example: self-observability
        name:
          type: string
          example: Self Observability
        enabled:
          type: boolean
        status:
          $ref: '#/components/schemas/ModuleStatus'
    Modules:
      type: object
      additionalProperties: false
      required: [modules]
      properties:
        modules:
          type: array
          items:
            $ref: '#/components/schemas/ModuleInventory'
    PlatformHealth:
      type: object
      additionalProperties: false
      required: [status, modules]
      properties:
        status:
          $ref: '#/components/schemas/ModuleStatus'
        modules:
          type: array
          items:
            $ref: '#/components/schemas/ModuleHealth'
    ServiceIdentity:
      type: object
      additionalProperties: false
      required: [name, namespace, environment]
      properties:
        name:
          type: string
          example: geordi-demo-service
        namespace:
          type: [string, 'null']
          example: geordi-demo
        environment:
          type: string
          example: development
    TimeRange:
      type: object
      additionalProperties: false
      required: [from, to]
      properties:
        from:
          type: string
          format: date-time
        to:
          type: string
          format: date-time
    OperationalMetric:
      type: string
      enum:
        - JVM_MEMORY_USED
        - JVM_CPU_UTILIZATION
        - JVM_THREAD_COUNT
        - JVM_GC_DURATION
        - HTTP_REQUEST_RATE
        - HTTP_REQUEST_COUNT
        - HTTP_REQUEST_LATENCY_P95
        - HTTP_ERROR_RATE
        - HTTP_ERROR_COUNT
    MetricUnit:
      type: string
      enum: [By, '1', '{thread}', s, '{request}/s', '{request}']
    MetricPoint:
      type: object
      additionalProperties: false
      required: [timestamp, value]
      properties:
        timestamp:
          type: string
          format: date-time
        value:
          type: number
          format: double
    MetricValue:
      type: object
      additionalProperties: false
      required: [metric, unit, value, timestamp]
      properties:
        metric:
          $ref: '#/components/schemas/OperationalMetric'
        unit:
          $ref: '#/components/schemas/MetricUnit'
        value:
          type: number
          format: double
        timestamp:
          type: string
          format: date-time
    MetricSeries:
      type: object
      additionalProperties: false
      required: [metric, unit, points]
      properties:
        metric:
          $ref: '#/components/schemas/OperationalMetric'
        unit:
          $ref: '#/components/schemas/MetricUnit'
        points:
          type: array
          items:
            $ref: '#/components/schemas/MetricPoint'
    MonitoredServices:
      type: object
      additionalProperties: false
      required: [services]
      properties:
        services:
          type: array
          items:
            $ref: '#/components/schemas/ServiceIdentity'
    MetricsOverview:
      type: object
      additionalProperties: false
      required: [service, range, values]
      properties:
        service:
          $ref: '#/components/schemas/ServiceIdentity'
        range:
          $ref: '#/components/schemas/TimeRange'
        values:
          type: array
          items:
            $ref: '#/components/schemas/MetricValue'
    MetricSeriesResult:
      type: object
      additionalProperties: false
      required: [service, range, series]
      properties:
        service:
          $ref: '#/components/schemas/ServiceIdentity'
        range:
          $ref: '#/components/schemas/TimeRange'
        series:
          type: array
          items:
            $ref: '#/components/schemas/MetricSeries'
    TraceId:
      type: string
      description: Lowercase 32-character hexadecimal OpenTelemetry trace ID.
      pattern: '^[0-9a-f]{32}$'
      example: 4bf92f3577b34da6a3ce929d0e0e4736
    SpanId:
      type: string
      description: Lowercase 16-character hexadecimal OpenTelemetry span ID.
      pattern: '^[0-9a-f]{16}$'
      example: 00f067aa0ba902b7
    TraceServiceIdentity:
      type: object
      additionalProperties: false
      required: [name, namespace, environment]
      properties:
        name:
          type: string
          minLength: 1
        namespace:
          type: [string, 'null']
        environment:
          type: [string, 'null']
    TraceServices:
      type: object
      additionalProperties: false
      required: [services]
      properties:
        services:
          type: array
          items:
            $ref: '#/components/schemas/TraceServiceIdentity'
    TraceSummary:
      type: object
      additionalProperties: false
      required: [traceId, rootSpanName, startTime, durationNanos, spanCount, error]
      properties:
        traceId:
          $ref: '#/components/schemas/TraceId'
        rootSpanName:
          type: string
        startTime:
          type: string
          format: date-time
        durationNanos:
          type: integer
          format: int64
          minimum: 0
        spanCount:
          type: integer
          minimum: 1
        error:
          type: boolean
    TraceSearchResult:
      type: object
      additionalProperties: false
      required: [service, range, traces]
      properties:
        service:
          $ref: '#/components/schemas/TraceServiceIdentity'
        range:
          $ref: '#/components/schemas/TimeRange'
        traces:
          type: array
          maxItems: 50
          items:
            $ref: '#/components/schemas/TraceSummary'
    SpanKind:
      type: string
      enum: [UNSPECIFIED, INTERNAL, SERVER, CLIENT, PRODUCER, CONSUMER]
    SpanStatus:
      type: string
      enum: [UNSET, OK, ERROR]
    HttpMetadata:
      type: object
      additionalProperties: false
      required: [requestMethod, route, path, responseStatusCode, serverAddress, serverPort]
      properties:
        requestMethod:
          type: [string, 'null']
        route:
          type: [string, 'null']
        path:
          type: [string, 'null']
        responseStatusCode:
          type: [integer, 'null']
          minimum: 100
          maximum: 599
        serverAddress:
          type: [string, 'null']
        serverPort:
          type: [integer, 'null']
          minimum: 1
          maximum: 65535
    TraceSpan:
      type: object
      additionalProperties: false
      required: [traceId, spanId, parentSpanId, name, service, telemetryOrigin, kind, status, startTime, startOffsetNanos, durationNanos, error, errorType, http]
      properties:
        traceId:
          $ref: '#/components/schemas/TraceId'
        spanId:
          $ref: '#/components/schemas/SpanId'
        parentSpanId:
          oneOf:
            - $ref: '#/components/schemas/SpanId'
            - type: 'null'
        name:
          type: string
        service:
          $ref: '#/components/schemas/TraceServiceIdentity'
        telemetryOrigin:
          type: [string, 'null']
        kind:
          $ref: '#/components/schemas/SpanKind'
        status:
          $ref: '#/components/schemas/SpanStatus'
        startTime:
          type: string
          format: date-time
        startOffsetNanos:
          type: integer
          format: int64
          minimum: 0
        durationNanos:
          type: integer
          format: int64
          minimum: 0
        error:
          type: boolean
        errorType:
          type: [string, 'null']
        http:
          oneOf:
            - $ref: '#/components/schemas/HttpMetadata'
            - type: 'null'
    TraceDetail:
      type: object
      additionalProperties: false
      required: [traceId, startTime, durationNanos, spanCount, error, spans]
      properties:
        traceId:
          $ref: '#/components/schemas/TraceId'
        startTime:
          type: string
          format: date-time
        durationNanos:
          type: integer
          format: int64
          minimum: 0
        spanCount:
          type: integer
          minimum: 1
        error:
          type: boolean
        spans:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/TraceSpan'
    Problem:
      type: object
      description: RFC 9457 problem details without provider-specific query text.
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
        detail:
          type: string